Moved shark fin plot to pure_colors.py.
