+++ colorpy/pure_colors.py
@@ -3,6 +3,11 @@ pure_colors.py - Some work-in-progress related to pure colors.
 
 Pure colors mean spectral lines and the 'purples'.
 
+Specialized plots:
+
+shark_fin_plot () -
+    Draw the 'shark fin' CIE chromaticity diagram of the pure spectral lines (plus purples) in xy space.
+
 This file is part of ColorPy.
 '''
 from __future__ import absolute_import
@@ -10,9 +15,10 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
+import io
 import math
 import numpy
-import io
+import matplotlib.pyplot as pyplot
 
 import colormodels
 import ciexyz
@@ -450,12 +456,117 @@ def perceptually_equal_spaced_color_plots ():
 
 #
 # 'Shark fin' chromaticity plot.
-# FIXME: That perhaps should move to this source file.
 #
 
 def shark_fin_plot ():
-    import plots
-    plots.shark_fin_plot()
+    '''Draw the 'shark fin' CIE chromaticity diagram of the pure spectral lines (plus purples) in xy space.'''
+    # get array of (approximate) colors for the boundary of the fin
+    xyz_list = get_num_pure_colors (
+        num_spect=2351, num_purple=200, brightness=1.0)
+
+    # get normalized colors
+    xy_list = xyz_list.copy()
+    (num_colors, num_cols) = xy_list.shape
+    for i in range (0, num_colors):
+        colormodels.xyz_normalize (xy_list [i])
+    # get phosphor colors and normalize
+    red   = colormodels.PhosphorRed
+    green = colormodels.PhosphorGreen
+    blue  = colormodels.PhosphorBlue
+    white = colormodels.PhosphorWhite
+    colormodels.xyz_normalize (red)
+    colormodels.xyz_normalize (green)
+    colormodels.xyz_normalize (blue)
+    colormodels.xyz_normalize (white)
+
+    def get_direc_to_white (xyz):
+        '''Get unit vector (xy plane) in direction of the white point.'''
+        direc = white - xyz
+        mag = math.hypot (direc [0], direc [1])
+        if mag != 0.0:
+            direc /= mag
+        return (direc[0], direc[1])
+
+    # plot
+    pyplot.clf ()
+
+    # draw best attempt at pure spectral colors on inner edge of shark fin
+    s = 0.025     # distance in xy plane towards white point
+    for i in range (0, len (xy_list)-1):
+        x0 = xy_list [i][0]
+        y0 = xy_list [i][1]
+        x1 = xy_list [i+1][0]
+        y1 = xy_list [i+1][1]
+        # get unit vectors in direction of white point
+        (dir_x0, dir_y0) = get_direc_to_white (xy_list [i])
+        (dir_x1, dir_y1) = get_direc_to_white (xy_list [i+1])
+        # polygon vertices
+        poly_x = [x0, x1, x1 + s*dir_x1, x0 + s*dir_x0]
+        poly_y = [y0, y1, y1 + s*dir_y1, y0 + s*dir_y0]
+        # draw (using full color, not normalized value)
+        color_string = colormodels.irgb_string_from_rgb (
+            colormodels.rgb_from_xyz (xyz_list [i]))
+        pyplot.fill (poly_x, poly_y, color_string, edgecolor=color_string)
+
+    # fill in the monitor gamut with true colors
+    def get_brightest_irgb_string (xyz):
+        '''Convert the xyz color to rgb, scale to maximum displayable brightness, and convert to a string.'''
+        rgb = colormodels.brightest_rgb_from_xyz (xyz)
+        color_string = colormodels.irgb_string_from_rgb (rgb)
+        return color_string
+
+    def fill_gamut_slice (v0, v1, v2):
+        '''Fill in a slice of the monitor gamut with the correct colors.'''
+        #num_s, num_t = 10, 10
+        #num_s, num_t = 25, 25
+        num_s, num_t = 50, 50
+        dv10 = v1 - v0
+        dv21 = v2 - v1
+        for i_s in range (num_s):
+            s_a = float (i_s)   / float (num_s)
+            s_b = float (i_s+1) / float (num_s)
+            for i_t in range (num_t):
+                t_a = float (i_t)   / float (num_t)
+                t_b = float (i_t+1) / float (num_t)
+                # vertex coords
+                v_aa = v0 + t_a * (dv10 + s_a * dv21)
+                v_ab = v0 + t_b * (dv10 + s_a * dv21)
+                v_ba = v0 + t_a * (dv10 + s_b * dv21)
+                v_bb = v0 + t_b * (dv10 + s_b * dv21)
+                # poly coords
+                poly_x = [v_aa [0], v_ba [0], v_bb [0], v_ab [0]]
+                poly_y = [v_aa [1], v_ba [1], v_bb [1], v_ab [1]]
+                # average color
+                avg = 0.25 * (v_aa + v_ab + v_ba + v_bb)
+                # convert to rgb and scale to maximum displayable brightness
+                color_string = get_brightest_irgb_string (avg)
+                pyplot.fill (poly_x, poly_y, color_string, edgecolor=color_string)
+    fill_gamut_slice (white, blue,  green)
+    fill_gamut_slice (white, green, red)
+    fill_gamut_slice (white, red,   blue)
+
+    # draw the curve of the xy values of the spectral lines and purples
+    pyplot.plot (xy_list [:,0], xy_list [:,1], color='#808080', linewidth=3.0)
+    # draw monitor gamut and white point
+    pyplot.plot ([red  [0], green[0]], [red  [1], green[1]], 'o-', color='k')
+    pyplot.plot ([green[0], blue [0]], [green[1], blue [1]], 'o-', color='k')
+    pyplot.plot ([blue [0], red  [0]], [blue [1], red  [1]], 'o-', color='k')
+    pyplot.plot ([white[0], white[0]], [white[1], white[1]], 'o-', color='k')
+    # label phosphors
+    dx = 0.01
+    dy = 0.01
+    pyplot.text (red   [0] + dx, red   [1], 'Red',   ha='left',   va='center')
+    pyplot.text (green [0], green [1] + dy, 'Green', ha='center', va='bottom')
+    pyplot.text (blue  [0] - dx, blue  [1], 'Blue',  ha='right',  va='center')
+    pyplot.text (white [0], white [1] + dy, 'White', ha='center', va='bottom')
+    # titles etc
+    pyplot.axis ([0.0, 0.85, 0.0, 0.85])
+    pyplot.xlabel (r'CIE $x$')
+    pyplot.ylabel (r'CIE $y$')
+    pyplot.title (r'CIE Chromaticity Diagram')
+    filename = 'ChromaticityDiagram'
+    # Save.
+    plots.plot_save (filename)
 
 #
 # Main.
